move the iteration required for select/operate/etc into collections module
