+++ dnp3/src/lib.rs
@@ -39,7 +39,7 @@
 //!
 //!     // In a real application, use the handle to make requests. Measurement data
 //!     // comes back via the handler specified when creating the association. See
-//!     // the provided examples for more details.
+//!     // the provided examples for more control.
 //!     tokio::time::delay_for(Duration::from_secs(60)).await;
 //!     Ok(())
 //! }
