+++ dnp3/src/outstation/control/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod collection;
+pub(crate) mod control_type;
+pub(crate) mod prefix;
