--- dnp3/src/outstation/details/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod control;
-pub(crate) mod prefix;
